requirements.txt
async-timeout==4.0.3
asyncpg==0.29.0
blinker==1.8.2
cachetools==5.3.3
cairocffi==1.7.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.2
et-xmlfile==1.1.0
Flask==3.0.3
Flask-WTF==1.2.1
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
gspread==6.1.2
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
marisa-trie==1.2.0
markup==0.2.2
MarkupSafe==2.1.5
mmh3==4.1.0
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.4
peewee==3.17.5
proto-plus==1.23.0
protobuf==4.25.3
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pygtrie==2.5.0
pyparsing==3.1.2
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.19.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.31
StrEnum==0.4.15
telebot==0.0.5
TgCrypto==1.2.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
wtf==0.1
WTForms==3.1.2
zipp==3.19.2
